+++ EfSamuariDomain/Entities/Moves/Punch.cs
@@ -4,7 +4,7 @@ namespace EfSamuariDomain.Entities.Moves
 {
     public class Punch : IOffensiveMove
     {
-        public float Damage { get; set; }
+        public double Damage { get; set; }
 
 
     }
